refactor: Generate receiving modal
Refactored generate eceiving modal to seperate out the info and form sections into their own components
