+++ src/components/GenerateReceivingModal/GenerateReceivingModalInfo/GenerateReceivingModalInfo.css
