+++ src/components/GenerateReceivingModal/GenerateReceivingModalInfo/index.js
@@ -1,0 +1,1 @@
+export { default } from './GenerateReceivingModalInfo';
